+++ megatron/layers/__init__.py
@@ -3,11 +3,13 @@ from . import common
 from . import image
 from . import missing
 from . import numeric
+from . import shaping
 from . import text
 from . import wrappers
 from .common import *
 from .image import *
 from .missing import *
 from .numeric import *
+from .shaping import *
 from .text import *
 from .wrappers import *