+++ megatron/utils/helpers.py
@@ -15,3 +15,7 @@ def safe_divide(X1, X2, impute=0):
     """
     impute_array = np.ones_like(X1) * impute
     return np.divide(X1.astype(np.float16), X2, out=impute_array.astype(np.float16), where=X2!=0)
+
+
+def vectorize(func):
+    return np.vectorize(func)